requirements.txt
python-dotenv
transformers
torch
torchvision
accelerate
huggingface-hub
langchain-groq
langchain-core
streamlit
requests
Pillow
gtts